+++ nata-monkey/src/main/java/com/nata/monkeys/MonkeyType.java
@@ -1,0 +1,13 @@
+package com.nata.monkeys;
+
+/**
+ * Author: Calvin Meng
+ * Blog: mclspace.com  Email: rdmclin2@gamil.com
+ * Update: 2016-04-25 12:14
+ */
+public class MonkeyType {
+    public static final String DFS= "dfs";
+    public static final String RM= "rm";
+    public static final String QLM= "qlm";
+    public static final String PRM= "prm";
+}