+++ nata-monkey/src/main/java/com/nata/server/Runner.java
@@ -14,21 +14,14 @@ public class Runner  extends Thread{
     AbstractMonkey monkey;
 
     public Runner(Config config){
-        AdbDevice device = new AdbDevice();
-        String appName = config.getApp_name();
-        String deviceId = config.getDevice_id();
-        int actionCount = config.getAction_count();
-        String packageName = config.getPackage_name();
-        String activityName = config.getActivity_name();
         String algorithm = config.getAlgorithm();
-        LogUtil.info("[Algorithm]: " + algorithm + " [App name]: " + appName + " [Device]: " + deviceId);
 
         switch (algorithm){
-            case "dfs" : monkey = new DfsMonkey(actionCount,packageName,activityName,device);break;
-            case "prm" : monkey = new PureRandomMonkey(actionCount,packageName,activityName,device);break;
-            case "rm"  : monkey = new RandomMonkey(actionCount,packageName,activityName,device);break;
-            case "qlm" : monkey = new QLearningMonkey(actionCount,packageName,activityName,device);break;
-            default    : monkey = new DfsMonkey(actionCount,packageName,activityName,device);break;
+            case "dfs" : monkey = new DfsMonkey(config);break;
+            case "prm" : monkey = new PureRandomMonkey(config);break;
+            case "rm"  : monkey = new RandomMonkey(config);break;
+            case "qlm" : monkey = new QLearningMonkey(config);break;
+            default    : monkey = new DfsMonkey(config);break;
         }
     }